+++ indico/modules/designer/blueprint.py
@@ -11,7 +11,8 @@ from indico.modules.designer.controllers import (RHAddCategoryTemplate, RHAddEve
                                                  RHCloneEventTemplate, RHDeleteDesignerTemplate,
                                                  RHDownloadTemplateImage, RHEditDesignerTemplate, RHGetTemplateData,
                                                  RHListBacksideTemplates, RHListCategoryTemplates, RHListEventTemplates,
-                                                 RHToggleTemplateDefaultOnCategory, RHUploadBackgroundImage)
+                                                 RHToggleBadgeDefaultOnCategory, RHToggleTemplateDefaultOnCategory,
+                                                 RHUploadBackgroundImage)
 from indico.util.caching import memoize
 from indico.web.flask.util import make_view_func
 from indico.web.flask.wrappers import IndicoBlueprint
@@ -33,6 +34,8 @@ def _dispatch(event_rh, category_rh):
 
 _bp.add_url_rule('/category/<int:category_id>/manage/designer/<int:template_id>/toggle-default',
                  'toggle_category_default', RHToggleTemplateDefaultOnCategory, methods=('POST',))
+_bp.add_url_rule('/category/<int:category_id>/manage/designer/<int:template_id>/toggle-default-badge',
+                 'toggle_category_default_badge', RHToggleBadgeDefaultOnCategory, methods=('POST',))
 
 
 for object_type in ('event', 'category'):