+++ indico/modules/designer/models/templates.py
@@ -112,6 +112,7 @@ class DesignerTemplate(db.Model):
 
     # relationship backrefs:
     # - backside_template_of (DesignerTemplate.backside_template)
+    # - default_badge_template_of (Category.default_badge_template)
     # - default_ticket_template_of (Category.default_ticket_template)
     # - images (DesignerImageFile.template)
     # - ticket_for_regforms (RegistrationForm.ticket_template)