+++ CHANGES.rst
@@ -108,6 +108,8 @@ Improvements
   other news items (:pr:`4587`)
 - Allow registrants to withdraw their application (:issue:`2715`, :pr:`4166`,
   thanks :user:`brabemi` & :user:`omegak`)
+- Allow choosing a default badge in categories (:pr:`4574`, thanks
+  :user:`omegak`)
 
 Bugfixes
 ^^^^^^^^